Fix mcrypt compatibility PHP 7.2
